% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{detect_insertion}
\alias{detect_insertion}
\title{detect_insertion}
\usage{
detect_insertion(ref_sequence, different_sequence, length_comparison = 10)
}
\arguments{
\item{ref_sequence}{Character vector whose elements are the bases of a DNA
sequence to use as reference.}

\item{different_sequence}{Character vector whose elements are the bases of a
DNA sequence different from the reference.}

\item{length_comparison}{Integer number. Number of bases to consider for the
comparison between the two DNA sequences in order to detect and remove
insertions in the non-reference sequence.}
}
\value{
Character vector of the different_sequence with length equal to
ref_sequence, after having removed the insertions.
}
\description{
detect_insertion
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
