% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{filter_bases}
\alias{filter_bases}
\title{filter_bases}
\usage{
filter_bases(heteroplasmy_matrix, min_heteroplasmy, min_cells, index = NULL)
}
\arguments{
\item{heteroplasmy_matrix}{Matrix with heteroplasmy values obtained from
\emph{get_heteroplasmy}.}

\item{min_heteroplasmy}{Numeric value.}

\item{min_cells}{Numeric value.}

\item{index}{index returned by \emph{get_heteroplasmy}. Default=NULL.}
}
\value{
Character vector of bases that have an heteroplasmy greater than
min_heteroplasmy in more than min_cells.
}
\description{
filter_bases
}
\author{
Gabriele Lubatti <gabriele.lubatti@helmholtz-muenchen.de>
}
