% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_tracing.R
\name{heatmap_plot}
\alias{heatmap_plot}
\title{heatmap_plot}
\usage{
heatmap_plot(
  marker_plot,
  marker_plot_plot,
  new_classification,
  old_classification,
  res_ang_sqrt,
  cluster_columns = F,
  cluster_rows = T,
  name_legend
)
}
\arguments{
\item{marker_plot}{Character vector. Samples to select for showing in the
heatmap.}

\item{marker_plot_plot}{Character vector. Selected samples will be labelled
with the elements of this vector in the heatmap.}

\item{new_classification}{Character vector with length equal to selected
samples,corresponding to the clustering analysis based on heteroplasmy
values. See \emph{clustering_dist_ang} for more info.}

\item{old_classification}{Character vector with length equal to selected
samples,corresponding to the partition available before clustering on
heteroplasmy values.}

\item{res_ang_sqrt}{Distance matrix obtained from \emph{clustering_dist_ang}}

\item{cluster_columns}{Logical. Parameter for cluster_columns argument of
the function \emph{Heatmap} in the package \emph{ComplexHeatmap}}

\item{cluster_rows}{Logical. Parameter for cluster_rows argument of the
function \emph{Heatmap}}

\item{name_legend}{Character value.Parameter for name argument of the
function \emph{Heatmap}}
}
\value{
Heatmap plot produced by function \emph{Heatmap}
}
\description{
heatmap_plot
}
\seealso{
[https://www.rdocumentation.org/packages/ComplexHeatmap/versions/1.10.2/topics/Heatmap]
}
\author{
Gabriele Lubatti <gabriele.lubatti@helmholtz-muenchen.de>
}
