% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_base_coverage}
\alias{plot_base_coverage}
\title{plot_base_coverage}
\usage{
plot_base_coverage(
  sum_matrix,
  sum_matrix_qc,
  selected_cells,
  interactive = FALSE,
  text_size = 10
)
}
\arguments{
\item{sum_matrix}{First element returned by the function
\emph{get_heteroplasmy}.}

\item{sum_matrix_qc}{Second element returned by the function
\emph{get_heteroplasmy}.}

\item{selected_cells}{Character vector with cells used fro plotting the
coverage.}

\item{interactive}{Logical. If TRUE an interactive plot is produced.}

\item{text_size}{Integer value specifying the size of the text in the plot.}
}
\value{
ggplot2 object (if interactive=FALSE) or plotly object (if
interactive=TRUE)
}
\description{
plot_base_coverage
}
\seealso{
[https://plotly.com/r/]
}
\author{
Gabriele Lubatti <gabriele.lubatti@helmholtz-muenchen.de>
}
