% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_batch}
\alias{plot_batch}
\title{plot_batch}
\usage{
plot_batch(
  position,
  heteroplasmy_matrix,
  batch,
  cluster,
  text_size,
  index = NULL
)
}
\arguments{
\item{position}{Character name of the base to plot.}

\item{heteroplasmy_matrix}{Matrix with heteroplasmy values obtained from
\emph{get_heteroplasmy}.}

\item{batch}{Vector of batch names,with length equal to n_row of
heteroplasmy_matrix.}

\item{cluster}{Vector of partition, with length equal to n_row of
heteroplasmy_matrix.}

\item{text_size}{Character specifying the size of the text for ggplot.}

\item{index}{index returned by \emph{get_heteroplasmy}. Default=NULL.}
}
\value{
ggplot object of the heteroplasmy level of a specific base across
samples divided according to batch.
}
\description{
plot_batch
}
\author{
Gabriele Lubatti <gabriele.lubatti@helmholtz-muenchen.de>
}
