% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{Plot_boxplot}
\alias{Plot_boxplot}
\title{Plot_boxplot}
\usage{
Plot_boxplot(
  distribution_1,
  distribution_2,
  label_1,
  label_2,
  name_x,
  name_y,
  name_title
)
}
\arguments{
\item{distribution_1, distribution_2}{Numeric vector}

\item{label_1}{Character vector of length equal to distribution_1}

\item{label_2}{Character vector of length equal to distribution_2}

\item{name_x}{Character name specifyng the xlab argument in ggplot.}

\item{name_y}{Character name specifyng the ylab argument in ggplot.}

\item{name_title}{Character name specifyng the ggplot2::ggtitle argument in ggplot.}
}
\value{
ggplot boxplot of the quantities specified by \emph{distribution_1}
and \emph{distribution_2}, divided by the conditions denoted by
\emph{label_1} and \emph{label_2}.
}
\description{
Plot_boxplot
}
\author{
Gabriele Lubatti <gabriele.lubatti@helmholtz-muenchen.de>
}
