% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_cells_coverage}
\alias{plot_cells_coverage}
\title{plot_cells_coverage}
\usage{
plot_cells_coverage(
  sum_matrix,
  cells_selected,
  cluster_ci,
  interactive = FALSE
)
}
\arguments{
\item{sum_matrix}{First element returned by the function
\emph{get_heteroplasmy}.}

\item{cells_selected}{Character vector of cells for which the coverage is
computed.}

\item{cluster_ci}{Character vector with partition information for cells
specified in \emph{cells_selected}}

\item{interactive}{Logical. If TRUE an interactive plot is produced.}
}
\value{
ggplot2 object (if interactive=FALSE) or plotly object (if
interactive=TRUE)
}
\description{
plot_cells_coverage
}
\seealso{
[https://plotly.com/r/]
}
\author{
Gabriele Lubatti <gabriele.lubatti@helmholtz-muenchen.de>
}
