% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_condition}
\alias{plot_condition}
\title{plot_condition}
\usage{
plot_condition(
  distribution_1,
  distribution_2,
  label_1,
  label_2,
  name_x,
  name_y,
  name_title
)
}
\arguments{
\item{distribution_1, distribution_2}{Numeric vector}

\item{label_1}{Character vector of length equal to distribution_1}

\item{label_2}{Character vector of length equal to distribution_2}

\item{name_x}{Character name specifyng the xlab argument in \emph{ggplot2}.}

\item{name_y}{Character name specifyng the ylab argument in ggplot2.}

\item{name_title}{Character name specifyng the ggtitle argument in \emph{ggplot2}.}
}
\value{
\emph{ggplot2} boxplot of the quantities specified by \emph{distribution_1}
and \emph{distribution_2}, separated by the conditions denoted by
\emph{label_1} and \emph{label_2}.
}
\description{
plot_condition
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
