% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_coordinate_heteroplasmy}
\alias{plot_coordinate_heteroplasmy}
\title{plot_coordinate_heteroplasmy}
\usage{
plot_coordinate_heteroplasmy(
  coordinate_dm,
  heteroplasmy_matrix_ci,
  index_ci,
  name_base
)
}
\arguments{
\item{coordinate_dm}{Dataframe whit samples on the rows and coordinates
names on the columns.}

\item{heteroplasmy_matrix_ci}{Third element returned by
\emph{get_heteroplasmy}.}

\item{index_ci}{Fifth element returned by \emph{get_heteroplasmy}.}

\item{name_base}{Character name for the base for which we want to plot the
expression level.}
}
\value{
ggplot2 object.
}
\description{
plot_coordinate_heteroplasmy
}
\author{
Gabriele Lubatti <gabriele.lubatti@helmholtz-muenchen.de>
}
