% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_correlation_bases}
\alias{plot_correlation_bases}
\title{plot_correlation_bases}
\usage{
plot_correlation_bases(bases_vector, index, heteroplasmy_matrix)
}
\arguments{
\item{bases_vector}{Character vector specifying the bases for which the
spearman correlation across samples is computed.}

\item{index}{Fifth element returned by \emph{get_heteroplasmy}.}

\item{heteroplasmy_matrix}{Third element returned by \emph{get_heteroplasmy}.}
}
\value{
Heatmap plot produced by function \emph{Heatmap}
}
\description{
plot_correlation_bases
}
\seealso{
\url{https://www.rdocumentation.org/packages/ComplexHeatmap/versions/1.10.2/topics/Heatmap}
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
