% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_tracing.R
\name{Plot_distance_matrix}
\alias{Plot_distance_matrix}
\title{Plot_distance_matrix}
\usage{
Plot_distance_matrix(dist_matrix_sc, cluster)
}
\arguments{
\item{dist_matrix_sc}{Distance matrix returned by function
\emph{clustering_dist_ang} (second element of the output).}

\item{cluster}{Character vector.Can be one of the two partitions returned by
function \emph{clustering_dist_ang} (first element of the output).}
}
\value{
Heatmap plot produced by function \emph{Heatmap}
}
\description{
Plot_distance_matrix
}
\seealso{
[https://www.rdocumentation.org/packages/ComplexHeatmap/versions/1.10.2/topics/Heatmap]
}
\author{
Gabriele Lubatti <gabriele.lubatti@helmholtz-muenchen.de>
}
