% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{Plot_distribution}
\alias{Plot_distribution}
\title{Plot_distribution}
\usage{
Plot_distribution(quantity_counts_cell, x_name, title_name)
}
\arguments{
\item{quantity_counts_cell}{numeric vector returned by
\emph{get_distribution.}}

\item{x_name}{Character name specifyng the xlab argument in ggplot.}

\item{title_name}{Character name specifyng the ggplot2::ggtitle argument in ggplot..}
}
\value{
ggplot density plot of the numeric vector quantity_counts_cell.
}
\description{
Plot_distribution
}
\author{
Gabriele Lubatti <gabriele.lubatti@helmholtz-muenchen.de>
}
