% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_tracing.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot_heatmap}
\usage{
plot_heatmap(
  new_classification,
  old_classification,
  dist_ang_matrix,
  cluster_columns = F,
  cluster_rows = T,
  name_legend
)
}
\arguments{
\item{new_classification}{Character vector.Second column of the dataframe
returned by function \emph{clustering_angular_distance} (first element of the
output).}

\item{old_classification}{Character vector. First column of the dataframe
returned by function \emph{clustering_angular_distance} (first element of the
output).}

\item{dist_ang_matrix}{Distance matrix obtained from \emph{clustering_angular_distance}
(second element of the output).}

\item{cluster_columns}{Logical. Parameter for cluster_columns argument of
the function \emph{Heatmap} in the package \emph{ComplexHeatmap}}

\item{cluster_rows}{Logical. Parameter for cluster_rows argument of the
function \emph{Heatmap}}

\item{name_legend}{Character value.Parameter for name argument of the
function \emph{Heatmap}}
}
\value{
Heatmap plot produced by function \emph{Heatmap}
}
\description{
plot_heatmap
}
\seealso{
\url{https://www.rdocumentation.org/packages/ComplexHeatmap/versions/1.10.2/topics/Heatmap}
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
