% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_heteroplasmy}
\alias{plot_heteroplasmy}
\title{plot_heteroplasmy}
\usage{
plot_heteroplasmy(position, heteroplasmy_matrix, cluster, index = NULL)
}
\arguments{
\item{position}{Character name of the base to plot.}

\item{heteroplasmy_matrix}{Matrix with heteroplasmy values obtained from
\emph{get_heteroplasmy}.}

\item{cluster}{Charachter vector specifying a partition of the samples, with
length equal to n_row of heteroplasmy_matrix.}

\item{index}{index returned by \emph{get_heteroplasmy}. Default=NULL.}
}
\value{
ggplot object of the heteroplasmy level of a specific base across
samples divided according to cluster.
}
\description{
plot_heteroplasmy
}
\author{
Gabriele Lubatti <gabriele.lubatti@helmholtz-muenchen.de>
}
