% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_heteroplasmy_variability}
\alias{plot_heteroplasmy_variability}
\title{plot_heteroplasmy_variability}
\usage{
plot_heteroplasmy_variability(
  heteroplasmy_matrix,
  cluster,
  threshold = 0.1,
  frac = FALSE,
  index
)
}
\arguments{
\item{heteroplasmy_matrix}{Third element returned by \emph{get_heteroplasmy}.}

\item{cluster}{Vector specifying a partition of the samples.}

\item{threshold}{Numeric value.}

\item{frac}{Logical. If FALSE the absolute number of cells that have at
least one base with heteroplasmy above \emph{threshold} are shown separated by
\emph{cluster}. If TRUE, then the fraction of cells are shown.}

\item{index}{Fifth element returned by \emph{get_heteroplasmy}.}
}
\value{
\emph{ggplot2} object
}
\description{
plot_heteroplasmy_variability
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
