% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_spider_chart}
\alias{plot_spider_chart}
\title{plot_spider_chart}
\usage{
plot_spider_chart(name_base, cluster_ci, heteroplasmy_matrix_ci, index_ci)
}
\arguments{
\item{name_base}{Character name specyfing the base for which the spider
chart is computed.}

\item{cluster_ci}{Character vector specyfing the partition for samples in
\emph{heteroplasmy_matrix_ci}.}

\item{heteroplasmy_matrix_ci}{Third element returned by
\emph{get_heteroplasmy}.}

\item{index_ci}{Fifth element returned by \emph{get_heteroplasmy}.}
}
\value{
radarchart plot produced by function \emph{radarchart}.
}
\description{
plot_spider_chart
}
\seealso{
[https://rdrr.io/cran/fmsb/man/radarchart.html]
}
\author{
Gabriele Lubatti <gabriele.lubatti@helmholtz-muenchen.de>
}
