% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{filter_bases}
\alias{filter_bases}
\title{filter_bases}
\usage{
filter_bases(heteroplasmy_matrix, min_heteroplasmy, min_cells, index = NULL)
}
\arguments{
\item{heteroplasmy_matrix}{Third element returned by \emph{get_heteroplasmy}.}

\item{min_heteroplasmy}{Numeric value.}

\item{min_cells}{Numeric value.}

\item{index}{Fifth element returned by \emph{get_heteroplasmy}.}
}
\value{
Character vector of bases that have an heteroplasmy greater than
\emph{min_heteroplasmy} in more than \emph{min_cells}.
}
\description{
filter_bases
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
