% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{get_distribution}
\alias{get_distribution}
\title{get_distribution}
\usage{
get_distribution(matrix, FUNCTION, index = NULL)
}
\arguments{
\item{matrix}{A matrix with samples on the rows and bases on the columns.}

\item{FUNCTION}{A character specifyng the function to be applied on each
column of \emph{matrix}. The possible values are:
\emph{mean},\emph{max},\emph{min},\emph{median} and \emph{sum}.}

\item{index}{index returned by \emph{get_heteroplasmy}. Default=NULL.}
}
\value{
It returns a vector with length equal to n_col of \emph{matrix}
where each element contains the result of the operation defined by
\emph{FUNCTION}.
}
\description{
get_distribution
}
\author{
Gabriele Lubatti <gabriele.lubatti@helmholtz-muenchen.de>
}
