% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_allele_frequency}
\alias{plot_allele_frequency}
\title{plot_allele_frequency}
\usage{
plot_allele_frequency(
  position,
  heteroplasmy_matrix,
  allele_matrix,
  cluster,
  names_allele_qc,
  names_position_qc,
  size_text,
  index = NULL
)
}
\arguments{
\item{position}{Character name of the base to plot.}

\item{heteroplasmy_matrix}{Matrix with heteroplasmy values obtained from
\emph{get_heteroplasmy}.}

\item{allele_matrix}{Matrix with allele frequencies values obtained from
\emph{get_heteroplasmy}.}

\item{cluster}{Characher vector specifying a partition of the samples, with
length equal to n_row of allele_matrix.}

\item{names_allele_qc}{Character vector with length equal to n_col of
allele_matrix. Each element specifies the name of the base and the allele.}

\item{names_position_qc}{Character vector with length equal to n_col of
allele_matrix. Each element specifies the name of the base.}

\item{size_text}{Character specifying the size of the text for ggplot
\emph{grid.arrange})}

\item{index}{index returned by \emph{get_heteroplasmy}. Default=NULL.}
}
\value{
\emph{grid.arrange} plot of allele frequencies of a specific base
across samples divided according to cluster.
}
\description{
plot_allele_frequency
}
\seealso{
[https://cran.r-project.org/web/packages/gridExtra/vignettes/arrangeGrob.html]
}
\author{
Gabriele Lubatti <gabriele.lubatti@helmholtz-muenchen.de>
}
