% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_coordinate_cluster}
\alias{plot_coordinate_cluster}
\title{plot_coordinate_cluster}
\usage{
plot_coordinate_cluster(coordinate_dm, cluster)
}
\arguments{
\item{coordinate_dm}{Dataframe whit samples on the rows and coordinates
names on the columns.}

\item{cluster}{Vector specifying a partition of the samples.}
}
\value{
\emph{ggplot2} object.
}
\description{
plot_coordinate_cluster
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
