% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_coordinate_heteroplasmy}
\alias{plot_coordinate_heteroplasmy}
\title{plot_coordinate_heteroplasmy}
\usage{
plot_coordinate_heteroplasmy(
  coordinate_dm,
  heteroplasmy_matrix,
  index,
  name_base
)
}
\arguments{
\item{coordinate_dm}{Dataframe whit samples on the rows and coordinates
names on the columns.}

\item{heteroplasmy_matrix}{Third element returned by \emph{get_heteroplasmy}.}

\item{index}{Fifth element returned by \emph{get_heteroplasmy}.}

\item{name_base}{Character name specyfing the base.}
}
\value{
\emph{ggplot2} object.
}
\description{
plot_coordinate_heteroplasmy
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
