% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_correlation_bases}
\alias{plot_correlation_bases}
\title{plot_correlation_bases}
\usage{
plot_correlation_bases(bases_example, index_ci, heteroplasmy_matrix_ci)
}
\arguments{
\item{bases_example}{Character vector specyfing the bases for which the
spearman correlation across samples is computed}

\item{index_ci}{Fifth element returned by \emph{get_heteroplasmy}.}

\item{heteroplasmy_matrix_ci}{Third element returned by
\emph{get_heteroplasmy}.}
}
\value{
Heatmap plot produced by function \emph{Heatmap}
}
\description{
plot_correlation_bases
}
\seealso{
[https://www.rdocumentation.org/packages/ComplexHeatmap/versions/1.10.2/topics/Heatmap]
}
\author{
Gabriele Lubatti <gabriele.lubatti@helmholtz-muenchen.de>
}
