% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_tracing.R
\name{plot_distance_matrix}
\alias{plot_distance_matrix}
\title{plot_distance_matrix}
\usage{
plot_distance_matrix(dist_ang_matrix, cluster)
}
\arguments{
\item{dist_ang_matrix}{Distance matrix obtained from \emph{clustering_angular_distance}
(second element of the output).}

\item{cluster}{Vector.Can be one of the two partitions returned by
function \emph{clustering_angular_distance} (first element of the output).}
}
\value{
Heatmap plot produced by function \emph{Heatmap}
}
\description{
plot_distance_matrix
}
\seealso{
\url{https://www.rdocumentation.org/packages/ComplexHeatmap/versions/1.10.2/topics/Heatmap}
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
