% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_distribution}
\alias{plot_distribution}
\title{plot_distribution}
\usage{
plot_distribution(quantity_counts_cell, name_x, name_title)
}
\arguments{
\item{quantity_counts_cell}{Vector returned by
\emph{get_distribution}}

\item{name_x}{Character name specifyng the xlab argument in \emph{ggplot2}.}

\item{name_title}{Character name specifyng the ggtitle argument in \emph{ggplot2}.}
}
\value{
\emph{ggplot2} density plot of the Vector quantity_counts_cell.
}
\description{
plot_distribution
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
