% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_heteroplasmy}
\alias{plot_heteroplasmy}
\title{plot_heteroplasmy}
\usage{
plot_heteroplasmy(position, heteroplasmy_matrix, cluster, index)
}
\arguments{
\item{position}{Character name of the base to plot.}

\item{heteroplasmy_matrix}{Third element returned by \emph{get_heteroplasmy}.}

\item{cluster}{Vector specifying a partition of the samples.}

\item{index}{Fifth element returned by \emph{get_heteroplasmy}.}
}
\value{
ggplot object of the heteroplasmy level of a specific base across
samples divided according to cluster.
}
\description{
plot_heteroplasmy
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
