% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_spider_chart}
\alias{plot_spider_chart}
\title{plot_spider_chart}
\usage{
plot_spider_chart(name_base, cluster, heteroplasmy_matrix, index)
}
\arguments{
\item{name_base}{Character name specifying the base.}

\item{cluster}{Vector specifying a partition of the samples.}

\item{heteroplasmy_matrix}{Third element returned by \emph{get_heteroplasmy}.}

\item{index}{Fifth element returned by \emph{get_heteroplasmy}.}
}
\value{
radarchart plot produced by function \emph{radarchart}.
}
\description{
plot_spider_chart
}
\seealso{
\url{https://rdrr.io/cran/fmsb/man/radarchart.html}
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
